Renamed SKSE plugin class and added mod files
Added mod files to match the directory structure of Skyrim, so that the
Data folder can just be copied into the Skyrim folder
