--- SkyrimTestScript.psc
@@ -1,11 +1,0 @@
-ScriptName SkyrimTestScript extends Quest
-
-FormList Property quests auto
-
-Event OnInit()
-	Quest questToBeChecked = quests.GetAt(5) as Quest
-	string questID = questToBeChecked.GetId()
-	string questname = questToBeChecked.GetName()
-	Debug.Notification(questID)
-	Debug.Notification(questname)
-EndEvent
