--- skybitica/skybitica/skybitica/MyPlugin.h
@@ -1,8 +1,0 @@
-#include "skse/PapyrusNativeFunctions.h"
-
-namespace MyPluginNamespace
-{
-	float MyTest(StaticFunctionTag *base);
-
-	bool RegisterFuncs(VMClassRegistry* registry);
-}
