+++ skybitica/skybitica/skybitica/Skybitica.h
@@ -1,0 +1,8 @@
+#include "skse/PapyrusNativeFunctions.h"
+
+namespace SkybiticaNamespace
+{
+	float AddQuest(StaticFunctionTag *base);
+
+	bool RegisterFuncs(VMClassRegistry* registry);
+}
