+++ skybitica/skybitica/skybitica/main.cpp
@@ -2,7 +2,7 @@
 #include "skse/skse_version.h"	// What version of SKSE is running?
 #include <shlobj.h>				// CSIDL_MYCODUMENTS
 
-#include "MyPlugin.h"
+#include "Skybitica.h"
 
 static PluginHandle					g_pluginHandle = kPluginHandle_Invalid;
 static SKSEPapyrusInterface         * g_papyrus = NULL;
@@ -10,15 +10,15 @@ static SKSEPapyrusInterface         * g_papyrus = NULL;
 extern "C"	{
 
 	bool SKSEPlugin_Query(const SKSEInterface * skse, PluginInfo * info)	{	// Called by SKSE to learn about this plugin and check that it's safe to load it
-		gLog.OpenRelative(CSIDL_MYDOCUMENTS, "\\My Games\\Skyrim\\SKSE\\MyPluginScript.log");
+		gLog.OpenRelative(CSIDL_MYDOCUMENTS, "\\My Games\\Skyrim\\SKSE\\Skybitica.log");
 		gLog.SetPrintLevel(IDebugLog::kLevel_Error);
 		gLog.SetLogLevel(IDebugLog::kLevel_DebugMessage);
 
-		_MESSAGE("MyPluginScript");
+		_MESSAGE("Skybitica");
 
 		// populate info structure
 		info->infoVersion =	PluginInfo::kInfoVersion;
-		info->name =		"MyPluginScript";
+		info->name =		"Skybitica";
 		info->version =		1;
 
 		// store plugin handle so we can identify ourselves later
@@ -45,12 +45,12 @@ extern "C"	{
 	}
 
 	bool SKSEPlugin_Load(const SKSEInterface * skse)	{	// Called by SKSE to load this plugin
-		_MESSAGE("MyScriptPlugin loaded");
+		_MESSAGE("Skybitica loaded");
 
 		g_papyrus = (SKSEPapyrusInterface *)skse->QueryInterface(kInterface_Papyrus);
 
 		//Check if the function registration was a success...
-		bool btest = g_papyrus->Register(MyPluginNamespace::RegisterFuncs);
+		bool btest = g_papyrus->Register(SkybiticaNamespace::RegisterFuncs);
 
 		if (btest) {
 			_MESSAGE("Register Succeeded");
