--- skybitica/skybitica/skybitica/scripts/MyPluginScript.psc
@@ -1,3 +1,0 @@
-scriptName MyPluginScript Hidden
-
-float Function MyTest() global native
