+++ skybitica/skybitica/skybitica/scripts/Skybitica.psc
@@ -1,0 +1,3 @@
+scriptName Skybitica Hidden
+
+float Function AddQuest() global native
