+++ skybitica/skybitica/skybitica/skybitica.vcxproj
@@ -118,11 +118,11 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="MyPlugin.cpp" />
+    <ClCompile Include="Skybitica.cpp" />
     <ClCompile Include="stdafx.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="MyPlugin.h" />
+    <ClInclude Include="Skybitica.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
