--- test/test/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="cpprestsdk.v140.windesktop.msvcstl.dyn.rt-dyn" version="2.8.0" targetFramework="native" />
-</packages>
