--- test/test/results.html
@@ -1,1 +1,0 @@
-{"text":"Test","type":"todo","id":"4f282e8b-3f6f-41c2-9079-eba2f218a119","notes":"","priority":1,"challenge":{},"attribute":"str","dateCreated":"2016-04-17T21:56:22.373Z","completed":false,"_id":"4f282e8b-3f6f-41c2-9079-eba2f218a119","value":0,"_editing":true,"_advanced":true}
